gk110/ir: Use an easier way to normalize input in rcp
The original way (of the blob) is correct, but hard to understand. The
resulting range is also irregular because of denorms. I reinvented a
equivalent way by setting exponent field directly to 0x3ff. The exponent
should be reextracted because of the denorm. However, instruction count
doesn't change.

This is part of effort to eliminate remaining magics in the rcp.